+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+- [turbulence] Start of the implementation for better turbulence fields.
 - [Shared, solvers] Improved turbulence model treatment.
 - [FoamResults] Addition of paraview option for results visualization in a seperate thread.
 - [snappyHexMesh, simpleFoam] Fixes for renaming boundary names.